Halo Combat Evolved Definitions Export First Pass

+++ Tags/BlamboozleLib/blamboozlelib-private-pch.h
@@ -14,7 +14,7 @@ struct tag_group
 {
 	union
 	{
-		uint32_t value;
+		unsigned long value;
 		char name[4];
 	};
 };
@@ -23,5 +23,6 @@ using btag_group = bigendian_type<tag_group>;
 
 #include "halo5/halo5_forge.h"
 #include "halo4/halo4.h"
+#include "halo1/halo1.h"
