+++ Tags/BlamboozleLib/halo4/h4_tag_enum.h
@@ -9,7 +9,7 @@ protected:
 public:
 	const char* name;
 	const char* filepath;
-	uint32_t line_number;
+	unsigned long line_number;
 	std::vector<const char*> options;
 
 	c_h4_tag_enum(const char* h4_data, const s_h4_tag_enum_definition* enum_definition);
