+++ Tags/BlamboozleLib/halo4/halo4.h
@@ -1,8 +1,8 @@
 #pragma once
 
 static constexpr uint64_t h4_base_address = 0x82000600;
-static constexpr uint32_t h4_tag_layout_table_address = 0x859D8350;
-static constexpr uint32_t h4_num_tag_layouts = 262;
+static constexpr unsigned long h4_tag_layout_table_address = 0x859D8350;
+static constexpr unsigned long h4_num_tag_layouts = 262;
 
 
 const IMAGE_SECTION_HEADER* relative_virtual_address_to_section(DWORD relative_virtual_address, const char* raw_image_data);
@@ -14,54 +14,54 @@ DWORD relative_virtual_address_to_relative_raw_address(DWORD relative_virtual_ad
 //#define __h4_aa_to_pointer(data, address) ((char*)(address))
 //#define __h4_pointer_to_va(data, pointer) (h4_base_address + static_cast<uintptr_t>((((char*)pointer) - ((char*)data))))
 //
-////#define h4_va_to_pa(data, address) static_cast<uint32_t>(address) - h4_base_address
+////#define h4_va_to_pa(data, address) static_cast<unsigned long>(address) - h4_base_address
 ////#define h4_va_to_pointer(data, address) (data + h4_va_to_pa(data, address))
 //
 //
-//inline uint32_t h4_va_to_pa(const char* data, uint32_t address)
+//inline unsigned long h4_va_to_pa(const char* data, unsigned long address)
 //{
 //	DWORD result = relative_virtual_address_to_relative_raw_address(address - h4_base_address, data, true);
 //	return result;
 //}
 
-const char* h4_va_to_pointer2(const char* data, uint32_t address);
-inline const char* h4_va_to_pointer(const char* data, uint32_t address)
+const char* h4_va_to_pointer2(const char* data, unsigned long address);
+inline const char* h4_va_to_pointer(const char* data, unsigned long address)
 {
 	if (address == 0) return 0;
-	uint32_t new_address = address;
+	unsigned long new_address = address;
 	const char* result = h4_va_to_pointer2(data, new_address);
 	ASSERT(result);
 	return result;
 }
-inline const char* h4_va_to_pointer_safe(const char* data, uint32_t address)
+inline const char* h4_va_to_pointer_safe(const char* data, unsigned long address)
 {
 	if (address == 0) return 0;
-	uint32_t new_address = address;
+	unsigned long new_address = address;
 	const char* result = h4_va_to_pointer2(data, new_address);
 	return result;
 }
 template<typename T>
-inline const T* h4_va_to_pointer(const char* data, uint32_t address)
+inline const T* h4_va_to_pointer(const char* data, unsigned long address)
 {
 	return reinterpret_cast<const T*>(h4_va_to_pointer(data, address));
 }
 template<typename T>
-inline const T* h4_va_to_pointer_safe(const char* data, uint32_t address)
+inline const T* h4_va_to_pointer_safe(const char* data, unsigned long address)
 {
 	return reinterpret_cast<const T*>(h4_va_to_pointer_safe(data, address));
 }
-//inline const char* h4_va_to_pointer(const char* data, uint32_t address)
+//inline const char* h4_va_to_pointer(const char* data, unsigned long address)
 //{
 //	if (address == 0) return nullptr;
-//	uint32_t pa = h4_va_to_pa(data, address);
+//	unsigned long pa = h4_va_to_pa(data, address);
 //	ASSERT(pa != ~DWORD());
 //	return data + pa;
 //}
 
-//inline const char* h4_va_to_pointer_safe(const char* data, uint32_t address)
+//inline const char* h4_va_to_pointer_safe(const char* data, unsigned long address)
 //{
 //	if (address == 0) return nullptr;
-//	uint32_t pa = h4_va_to_pa(data, address);
+//	unsigned long pa = h4_va_to_pa(data, address);
 //	if (pa != ~DWORD())
 //	{
 //		return data + pa;
